foreach: Port to former colon-separation syntax
As per discussions in #1449.
Other changes:- "UNI" substitution now corresponds to the unique part of the input, after removing common prefix and suffix. Previously it was the unique part of the -basename-, after removing the common -suffix- only; this to me was not a general solution, and the behaviour was not necessarily consistent with the name of the substitution.
- Raise an exception if the command string following the colon separator does not contain any instances of any of the substitutions, and therefore the exact same command would be executed for every input.
- Update foreach documentation page accordingly, including demonstration of -nthreads use both before and after the colon separator.
- run.killAll(): Fix attempting to iterate over a NoneType.

+++ lib/mrtrix3/run.py
@@ -411,15 +411,16 @@ def killAll(): #pylint: disable=unused-variable
   _processes = [ ]
   # Also destroy any remaining temporary files
   for tlist in _tempFiles:
-    for tpair in tlist:
-      for t in tpair:
-        if t:
-          try:
-            os.remove(t)
-          except OSError:
-            pass
-      tpair = [ None, None ]
-    tlist = [ ]
+    if tlist:
+      for tpair in tlist:
+        for t in tpair:
+          if t:
+            try:
+              os.remove(t)
+            except OSError:
+              pass
+        tpair = [ None, None ]
+      tlist = None
   _tempFiles = [ ]
   _lock.release()
